refactor(experimental): use GraphQL lang for schema
This PR is a complete makeover of how this library is using GraphQL.js.

This change pulls in `@graphql-tools/schema` to build GraphQL type definitions
from _actual_ GraphQL language syntax.

You can see this in action in any of the schema files, such as:

- `src/schema/account.ts`
- `src/schema/common/inputs.ts`
- `src/schema/transaction.ts`

Overall this code is much more readable, easier to organize and keep track of,
and easier to tree-shake.

The library's codebase will scale much easier with this layout.