refactor(experimental): build libraries at the highest ECMAScript target possible (#1877)
# Summary

We're building a library. We don't deploy directly on the target – our customers do. They should have the final say what ECMA target they want to run against (eg. ES2020, ES2015, et cetera).

It's wise, therefore, to compile our _library_ at the highest possible target (ie. ESNext) so as not to lose any features/information on the way into our customers' build systems.

The only exception is the IIFE bundle that we ship, which we compile to the `browserslist` target "`supports bigint and not dead`".

Thanks to @hyrious for [the suggestion](https://github.com/evanw/esbuild/issues/3500#issuecomment-1820046687).

# Test Plan

Spot check the compiled source in `packages/library-legacy-sham`

1. The browser/node/native bundles still have all of the private property syntax intact
2. The developement/production IIFE bundles have the private property syntax compiled to an implementation based on `WeakMap`