refactor(experimental): add read write functions to codecs-strings (#1885)
This PR continues to refactor the codecs libraries as described in PR #1865. This one focuses on `@solana/codecs-strings`.

Here, we can see the downside of having to compute the size beforehand for base X since, unless we add some caching, we end up having to encode twice: once to get the size of the buffer and once to fill it.