refactor(experimental): rename fixCodec and prefixCodec to fixCodecSize and prefixCodecSize (#2411)
This PR renames `fixCodec` to `fixCodecSize` and `prefixCodec` to `prefixCodecSize` [as discussed here](https://github.com/solana-labs/solana-web3.js/pull/2397#discussion_r1546923105).

Note that I created a new changeset describing the change for `fixCodec` but not for `prefixCodec` as this is not yet merged and I thought it'd be better to update the changeset created by a previous PR of this stack.