Add `replacementBlockhash` to the simulation return type, in preparation for Agave 2.0 (#2870)
# Summary

When you ask the simulator to `replaceRecentBlockhash`, you may like to know what the replacement blockhash was. Agave 2.0 will return it in the simulation result.

It would have been nice to _only_ include this property when `replaceRecentBlockhash` is explicitly `true`, but I would have had to duplicate all of the overloads for the purpose, yielding 32 in total.