`__DEV__` logic is now ‘NODE_ENV is not production’ (#2907)
# Summary

A few folks have been caught out _not_ having _any_ setting for `NODE_ENV`. In cases like this, the `__DEV__` flag will be `false` because `NODE_ENV` is not `'development'`.

It would be preferable for development to be the default when there is no `NODE_ENV`. This means it will be harder to get ‘into’ production mode (you have to set an env variable) but it will mean that people fall into the pit of success when debug functionality is what they're looking for.

# Test Plan

```
pnpm i
pnpm turbo compile:js
```

Observe that all of the `__DEV__` constants have been replaced with `process.env.NODE_ENV !== 'production'`